test_swap.vim: fix Test_swapfile_delete
Test_swapfile_delete test assumes that pid 1 is running with a different
user as the test, which is not true during the build where the build is run
in a separate pid namespace.

To fix the test, the PID is updated to 0xdead0000 which is PID 3735879680 so
it can be assured that this will be a non-existing one and the swap file
will be removed.